Initial commit of RunD convolutional model. Tensorflow and custom bitaccurate versions are working, but hte precompiled model version is removed
This does some flattening of the tensorflow input to deal with the new flattened input layer. Much of the precompiled model infrastructue was all pulled out as it began to complain about text file locations (which should not have been overwritten). The precompiled model infrastructure is being tested, but this contains functional bit accurate input.