Remove redundant CICADA emulation techniques and associated files
This is the first commit in a series of tests designed to make the CICADA emulator ready for integration into CMSSW. The TensorFlow accurate emulation method, using the CMSSW TensorFlow interface has been removed, as it is not hardware accurate. The method using the complete HLS4ML code and special compilation options has been removed for adding too many files, too many compiler exemptions, and not having a good version control solution. This repository compiles, and is expected to work, but has not yet been tested. Code formatting has been applied.